Create new snapshots directory
